Forget harder.

+++ geom.h
@@ -56,7 +56,8 @@ inline double bmm_geom_ballsurf(double const r, size_t d) {
 /// $J_d(r) = ((d - 1) / (d + 2)) m r^2$.
 __attribute__ ((__const__, __pure__))
 inline double bmm_geom_ballmoi(double const r, size_t const d) {
-  return d == 0 ? NAN : ((double) (d - 1) / (double) (d + 2)) * bmm_fp_sq(r);
+  return d == 0 ?
+    (double) NAN : ((double) (d - 1) / (double) (d + 2)) * bmm_fp_sq(r);
 }
 
 /// The call `bmm_geom_ballpmoi(r, d)` returns the moment of inertia
